Customer expected rate escalations now based on AEO2014 projections
In cashflow calculations, customers have a forecast of real $2011 rate
changes over the lifetime of the system. The previous value was a
constant 2.5%, which is much higher than historic changes in real terms.
Now, the expected rate is based on the average percent change over the
life of the project based on AEO regional rate projections (again, in
$2011). For years after 2040 (end of AEO projection), rate increases are
flat.

The effect of the change should be to decrease economic attractiveness
and deployment because future rates will look much more reasonable