Partial commit for leasing model
Partial commit for the leasing model. This version is a work in progress
and doesn't compile.

Remaining tasks:

- Decision to buy or lease made in scoe or generate_customer_bins. What
is the rationale for this?
-Logic for when leasing is available in a state
-Ensure the calc_cashflows, economics, and incentives are correct. By
this I mean--are they getting the correct monetizations as a commercial
entity? More fundamentally, is leasing priced accurately and how to
account for leasing cost of doing business vs economy of scale
- Completing max_market_share joins
- Completing calc_diffusion logic
- How to make to quantify that leasing is more attractive than buying