Misc changes to payback period and cashflows
Revised cash flow calculation s.t. down payment occurs in year 0.
Previously the down payment and the first year revenues and costs were
combined in the first (zero-index) column of the cashflow arrays

Revised payback period calculation reverting to previous algorithm,
which is better described as time-to-net-positive cashflow. In short
hand, it is the first period in which net (cumulative) cashflows are
positive for rest of the investment. This can produce questionably low
paybacks when the down payment fraction is very low.