Bug fix to DSIRE incentives
Wow, this was a big one! The value of the DSIRE incentives was being
overstated by a 2x value. When we calculate the value of dsire
incentives we calculate the value of each incentive_array_id, then group
by county and bin id, then sum over the incentive value for each unique
group. However with the addition of leasing, that means that grouping by
county_id, bin_id is insufficient, and we must group by county_id,
bin_id, AND business_model to find each unique customer.