Fix to 2.5kW issues for Wind Aggressive case
This should fix a few of the issues we saw in the Wind 'Aggressive'
case.

First, the 2.5kW now has the 20m option selected as default. A
consequence of this is that the 2.5kW model can't get cheaper by
shifting down from 30m to 20m as it could before

Second, fixes a bug in the combine_temporal function that should improve
performance and also prevent turbines from selecting resource at the 40m
height with 20m costs (this is true across all turbine sizes, not just
the 2.5 model).

As a consequence, the reference case now installs 9.1 GW, down from
9.3GW and the aggressive case installs 13.4 GW from 17.4 GW (bigger
change)