Calculate percent_excess_generation
Using the consumptio and generation profiles in rate_input_df for SAM
bill savings, calculate the percent of annual generation which is
exported to the grid. This will be material in NEM calculations as well
as in granting more options in applying curtailments. For now I have
implemented applying curtailments to either the  'net' generation e.g.
only generation exported to grid or to the 'gross' generation e.g. all
system generation.

Remaining work: need to implement appropriate options on the ReEDS side,
particularly, passing curtailment_method as well as implementing
penetration threshholds to begin enforcing curtailments.