Adds option to use CG's new TPO MMS curve
This hacks the model to read in a csv that will replace the max market
share curves ingested. As written, it will completely ignore the market
share settings in the input sheet and just use the 'NREL' curves in the
csv.

Once we are comfortable with using these results, it should be simple to
replace the 'NREL' rows of the max_market_share in the
diffusion_template